plugin-local-projectevaluator/python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
attrs==25.3.0
boto3==1.40.35
botocore==1.40.35
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
cryptography==46.0.0
dataclasses-json==0.6.7
fastapi==0.117.1
filelock==3.19.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.29
lxml==6.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.31.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.17.3
snowflake-snowpark-python==1.39.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tomlkit==0.13.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.36.0
wheel==0.45.1
yarl==1.20.1
zstandard==0.25.0
